pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = "created by team 32"
authors = ["ramanuja <es22btech11022@iith.ac.in>"]

[tool.poetry.dependencies]
python = "^3.10,<3.15"
boto3 = "^1.40.16"
botocore = "^1.40.16"
langchain = "^0.3.27"
langchain-community = "^0.3.27"
langchain-aws = "^0.2.31"
python-dotenv = "^1.1.1"
langchain-google-genai = "^2.1.9"
faiss-cpu = "^1.12.0"
sentence-transformers = "^5.1.0"



[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]


[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 